Revert "Lockscreen charging info: show decimal in battery temperature"
This reverts commit 77b344f7f56a91332ed90e6c80a83559a060c937.

+++ packages/SystemUI/src/com/android/systemui/statusbar/KeyguardIndicationController.java
@@ -187,7 +187,7 @@ public class KeyguardIndicationController {
     private int mChargingWattage;
     private int mBatteryLevel;
     private boolean mBatteryPresent = true;
-    private float mTemperature;
+    private int mTemperature;
     private long mChargingTimeRemaining;
     private String mBiometricErrorMessageToShowOnScreenOn;
     private final Set<Integer> mCoExFaceAcquisitionMsgIdsToShow;
